Move some inputs.py classes to builder/inputs.py
From nengo_loihi.inputs. These classes are only used as part of
the build process and should not be made alongside Block and
LoihiInput instances.
