+++ nengo_loihi/builder/connection.py
@@ -15,9 +15,10 @@ import numpy as np
 from nengo_loihi import conv
 from nengo_loihi.block import Axon, LoihiBlock, Probe, Synapse
 from nengo_loihi.builder.builder import Builder
+from nengo_loihi.builder.inputs import ChipReceiveNeurons
 from nengo_loihi.compat import (
     nengo_transforms, sample_transform, conn_solver)
-from nengo_loihi.inputs import ChipReceiveNeurons, LoihiInput
+from nengo_loihi.inputs import LoihiInput
 from nengo_loihi.neurons import loihi_rates
